+++ src/app/api/categories/route.js
@@ -3,8 +3,15 @@ import { NextResponse } from "next/server";
 
 export const GET = async (request) => {
     const categories = await db.category.findMany({
-        include: {
-            posts: true,
+        select: {
+            id: true,
+            name: true,
+            description: true,
+            _count: {
+                select: {
+                    posts: true
+                }
+            },
         },
     });
     return NextResponse.json(categories);